requirements.txt
streamlit
pillow
numpy
opencv-python-headless
tensorflow==2.8.0
ultralytics
